Add customisation of auto-gen'ed SQL record names
This adds an Aeson-inspired ability to configure how
`deriveEsqueletoRecord` generates the SQL record type, in particular
which names it generates for the record type itself and its fields.

This is because `DuplicateRecordFields` is a painful extension to use,
and avoiding it through module structuring is also painful. A nicer
solution for avoiding duplicate record fields is to simply not generate
them, and allow the user to specify how record fields should be
generated so they don't overlap.